refactor: udpate events
